use arch independent types for serialised values

+++ include/mapnik/json/extract_bounding_boxes_x3_config.hpp
@@ -67,11 +67,11 @@ struct extract_positions
 };
 
 using box_type = mapnik::box2d<double>;
-using boxes_type = std::vector<std::pair<box_type, std::pair<std::size_t, std::size_t>>>;
+using boxes_type = std::vector<std::pair<box_type, std::pair<std::uint64_t, std::uint64_t>>>;
 using callback_type = extract_positions<grammar::iterator_type, boxes_type>;
 
 using box_type_f = mapnik::box2d<float>;
-using boxes_type_f = std::vector<std::pair<box_type_f, std::pair<std::size_t, std::size_t>>>;
+using boxes_type_f = std::vector<std::pair<box_type_f, std::pair<std::uint64_t, std::uint64_t>>>;
 using callback_type_f = extract_positions<grammar::iterator_type, boxes_type_f>;
