+++ plugins/input/geojson/geojson_index_featureset.hpp
@@ -41,7 +41,7 @@
 
 class geojson_index_featureset : public mapnik::Featureset
 {
-    using value_type = std::pair<std::size_t, std::size_t>;
+    using value_type = std::pair<std::uint64_t, std::uint64_t>;
 public:
     geojson_index_featureset(std::string const& filename, mapnik::filter_in_box const& filter);
     virtual ~geojson_index_featureset();
