+++ utils/mapnik-index/process_csv_file.cpp
@@ -93,7 +93,7 @@ std::pair<bool,typename T::value_type::first_type> process_csv_file(T & boxes, s
 }
 
 using box_type = mapnik::box2d<float>;
-using item_type = std::pair<box_type, std::pair<std::size_t, std::size_t>>;
+using item_type = std::pair<box_type, std::pair<std::uint64_t, std::uint64_t>>;
 using boxes_type = std::vector<item_type>;
 template std::pair<bool,box_type> process_csv_file(boxes_type&, std::string const&, std::string const&, char, char);
