+++ utils/mapnik-index/process_geojson_file_x3.cpp
@@ -213,7 +213,7 @@ bool validate_geojson_feature(mapnik::json::geojson_value & value, Keys const& k
 };
 
 using box_type = mapnik::box2d<float>;
-using boxes_type = std::vector<std::pair<box_type, std::pair<std::size_t, std::size_t>>>;
+using boxes_type = std::vector<std::pair<box_type, std::pair<std::uint64_t, std::uint64_t>>>;
 using base_iterator_type = char const*;
 
 auto const& geojson_value = mapnik::json::geojson_grammar();
